remove vcenter package move filtering logic into template update README.md
